% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_airbnb.R
\name{mod_airbnb_ui}
\alias{mod_airbnb_ui}
\alias{mod_airbnb_server}
\title{mod_airbnb_ui and mod_airbnb_server}
\usage{
mod_airbnb_ui(id, dest)

mod_airbnb_server(input, output, session, dest)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
