% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_places.R
\name{mod_places_ui}
\alias{mod_places_ui}
\alias{mod_places_server}
\title{mod_places_ui and mod_places_server}
\usage{
mod_places_ui(id, dest)

mod_places_server(input, output, session, dest)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
