% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_statistics.R
\name{mod_statistics_ui}
\alias{mod_statistics_ui}
\alias{mod_statistics_server}
\title{mod_statistics_ui and mod_statistics_server}
\usage{
mod_statistics_ui(id, dest)

mod_statistics_server(input, output, session, dest)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
