% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_weather.R
\name{mod_weather_ui}
\alias{mod_weather_ui}
\alias{mod_weather_server}
\title{mod_weather_ui and mod_weather_server}
\usage{
mod_weather_ui(id, dest)

mod_weather_server(input, output, session, dest)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
